[EXPERIMENTAL] LiteDB chunk storage
I needed a quick and easy database engine... LiteDB is that. LiteDB is
far from amazing or powerful or ... capable or any other positive
attribute really at the moment, but after a bit of screwing around it
functions better than my old lazy filesystem. Faster too. This may be a
placeholder for some other DB system in the future, but it serves its
purpose for now.